feat: deleted old BridgeSupportProcessFundsMigrationTest and updated it with the new ProcessFundsMigration tests
